Replace copied Contactable with prepended module
Instead of maintaining a copy of Contactable in our code, we define a small
module that overrides just the method providing the default zip code
country. As we are overriding a method defined in the Contactable module,
we have to use prepend to insert our module in the beginning of the
ancestor chain, i.e., adding it before the original Contactable to the
function lookup order.