Hide main navigation items through permissions
We hid the items in the main navigation (column on the left) by copying the
navigation helper and overwriting the conditions to display the items.
Instead, let's use the permissions for that. The previously hidden items
would be visible to everyone by default, make them only accessible to admin
users for now. With that, we can drop the copied navigation helper from our
code base.