Remove subscriptions and events from groups
We don't use the subscriptions and events yet, so the tabs are empty and
useless at the moment. Hide them for now to avoid confused users. They are
hidden by setting the abilities so that the events and mailing lists tabs
on groups are only shown to users if they have a role with permissions on
the group.