Show medical pages belonging to person
All pages that belong to a person have a white background for their inner
sheet. This used to happen only on 'index' actions. Add a class attr to set
this display mode for all pages of a sheet.

We need to patch the active tab detection, so it matches the exact path
before the path prefix, e.g., '/groups/2/people/3/medical/edit' should not
match '/groups/2/people/3/' (and thereby highlight the first tab) if there
is an exact match available.