Show more years in birthday datepicker
Instead of just -10 to +10 years, show all years from 1900 to this year.
While any date with the current year will be invalid for participation
because of age restrictions, this makes sure the initially selected year in
the date picker is 2025 instead of 1900 (default is to select the first
when the current year is not in the range). This way, most participants
only need to scroll a few years back.