Allow to include ContractHelper directly in Person
Person has an attribute called payment_role and the ContractHelper had an
functions called payment_role. When you included ContractHelper in Person
to access the ul?, yp?, etc. functions, the attribute and the function
conflicted. To resolve this, rename the function in the ContractHelper and
change all occurrences. Also centralize the early payer check.

The PersonDecorator is needed to access current_roles_grouped. We now
create the PersonDecorator closer to where it is actually needed instead of
passing it around through many layers. However, creating a PersonDecorator
around a PersonDecorator is not a problem, so if I missed a spot where this
is created, it does no harm.