Improve performance of map
Loading the map took a long time because serializing all Persons into JSON
did include the tag_list for each Person. This happened because the
acts-as-taggable-on gem adds the tag_list as an attribute to the Person
model. These attributes are calculated when the Ruby object is created,
i.e., after the ActiveRecord .select where we correctly limit the data we
pull from the database.

We don't need the tags for the map at all. Therefore, exclude the tag_list
from the JSON objects to prevent loading tags for each person from the
database.