Do not select people as being in the same rdp group if the rdp group is NULL
Problem: When a person in a cluster has all rdp association fields
set to NULL, then all other people not having set any of the rdp
association fields are seen as being in the same group.

Solution: Exclude people where rdp_association_group is NULL from
being considered to be in the same rdp group as a person in the
cluster. This still allows some rdp association fields to be NULL, but
should resolve the most problematic cases without being a restriction.