Make pre notification db setup more flexible
Problem: We need to reference a direct debit payment info when
inserting a new direct debit pre-notification row. This way it is
complicated to change the sequence type between pre-notification and
finalization of the payment initiation.

Solution: Allow the creation of a direct debit pre-notification
without refering to a direct debit payment info. The creation of the
payment info entries can thus be delayed until we finalize the payment
initiation.